feat: add routes for routing to different role dashboards
