+++ components/layout/Navbar/NavbarButtonsMobile/NavbarButtonMobile/NavbarButtonMobile.tsx
@@ -2,10 +2,10 @@ import { FC } from "react";
 // Components
 import { MenuItem, Typography } from "@mui/material";
 // Helpers
-import { NavbarOption } from "@/helpers/navigation";
 import { userHasRole } from "@/helpers/roles";
 // Types
 import { User } from "@/types/users";
+import { NavbarOption } from "../../Navbar.types";
 
 type Props = {
   user?: User;
